Fix ceph-disk when running with dmcrypt on Ubuntu
On Ubuntu with dmcrypt, partprobe doesn't do
enough to reload all of the necessary device
metadata that's required for udev to properly
activate the disks. Specifically,
ID_PART_ENTRY_TYPE is not updated by partprobe.
partx -u does work properly, so add a new call
to update_partition() and tweak it slightly to
make the proper partx call.